SPACE-22: Fixed the useHazard tests to be useEncounter.
